pipeline/scripts/requirements.txt
biopython>=1.79
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
torch>=1.9.0
transformers>=4.12.0
umap-learn>=0.5.0
faiss-cpu>=1.7.0
hdbscan>=0.8.0
networkx>=2.6.0
ete3>=3.1.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
kaleido>=0.2.1

